test_unfold/requirements.txt
django
django-unfold
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
django-filter
mysqlclient
python-dotenv
payos